Fix compiler warnings for useless cast (#1175)
役に立たないキャストを行っているとgccに指摘されたため修正します。

gcc 12のレポート
```
../src/article/drawareabase.cpp:3189:64: warning: useless cast to type ‘int’ [-Wuseless-cast]
 3189 |                     else if( mode == LIVE_SCRMODE_STEADY ) y = ( int ) ( current_y + CONFIG::get_live_speed() );
      |                                                                ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../src/skeleton/edittreeview.cpp:271:46: warning: useless cast to type ‘class Gtk::TreePath’ [-Wuseless-cast]
  271 |             Gtk::TreeRow row = get_row( Gtk::TreePath( m_jump_path ) );
      |                                              ^~~~~~~~~~~~~~~~~~~~~~~
../src/environment.cpp:35:51: warning: useless cast to type ‘std::string’ {(snip)} [-Wuseless-cast]
   35 | std::string ENVIRONMENT::get_jdbbs(){ return std::string( JDBBS ); }
      |                                                   ^~~~~~~~~~~~~~~
../src/environment.cpp:36:54: warning: useless cast to type ‘std::string’ {(snip)} [-Wuseless-cast]
   36 | std::string ENVIRONMENT::get_jd2chlog(){ return std::string( JD2CHLOG ); }
      |                                                      ^~~~~~~~~~~~~~~~~~
../src/environment.cpp:35:51: warning: useless cast to type ‘std::string’ {(snip)} [-Wuseless-cast]
   35 | std::string ENVIRONMENT::get_jdbbs(){ return std::string( JDBBS ); }
      |                                                   ^~~~~~~~~~~~~~~
../src/environment.cpp:36:54: warning: useless cast to type ‘std::string’ {(snip)} [-Wuseless-cast]
   36 | std::string ENVIRONMENT::get_jd2chlog(){ return std::string( JD2CHLOG ); }
      |                                                      ^~~~~~~~~~~~~~~~~~
```

+++ src/article/drawareabase.cpp
@@ -3186,7 +3186,7 @@ void DrawAreaBase::exec_scroll()
 
                     const int mode = CONFIG::get_live_mode();
                     if( mode == LIVE_SCRMODE_VARIABLE ) y = ( int ) ( current_y + m_scrollinfo.live_speed );
-                    else if( mode == LIVE_SCRMODE_STEADY ) y = ( int ) ( current_y + CONFIG::get_live_speed() );
+                    else if( mode == LIVE_SCRMODE_STEADY ) y = current_y + CONFIG::get_live_speed();
                 }
 
                 // 行単位スクロール
